# metricvec: handle hash collision for labeled metrics