@stevvooe 2016-08-12
Take a peak https://github.com/prometheus/client_golang/pull/221/commits/9b464a48bdf43c423779a9dab1c3a1d5cb0c0d8c. It provides some further fixes around collision handling and greatly increases the unit test coverage.

Zero-alloc behavior is preserved at the cost of some performance.