@beorn7 2016-08-15
Thanks @stevvooe for the PR, thanks @grobie for the review. I'm back from my travels and will have a look ASAP. Zero-allocation and benchmarks look great.