@beorn7 2016-08-16
Looks good. Just a few naming and comment nits.

I'll try to dig out a fnv64a hash collision.