@beorn7 2016-08-16
`77kepQFQ8Kl` and `!0IC=VloaY` both have a hash value of 0. You can use them in the test as label values so that you don't need to inject the hashing function anymore.