@stevvooe 2016-08-16
> 77kepQFQ8Kl and !0IC=VloaY both have a hash value of 0. You can use them in the test as label values so that you don't need to inject the hashing function anymore.

Thanks for hunting this down, but this won't test the 2+N search case. That extra test caught a number of bugs in both the new and original implementation.