@beorn7 2016-08-16
More strings with hash value 0:

<pre>
=. hx"iX<;
QvXtM%gt;@Fp%
_"kWk=-v$c
</pre>

See http://www.isthe.com/chongo/tech/comp/fnv/