@beorn7 2016-08-17
Tests are fine. I'm merging this now, as I want to have it int the release planned for today.

Feel free to add a test with a true collision in a separate PR. I can also do it myself.

Thanks for your contribution.