@beorn7 2016-08-17
Only now I get it: the benchmarks in the commit description are superseeded by the following commit.

We indeed have to rework to have no allocations in the `WithLabelValues` case.
But don't worry, I'm already working on a solution.
(And I have also added a test with a "real" collision in #225.)

Thanks again.