@stevvooe 2016-08-17
@beorn7 Alright, thanks for taking the PR.

Let me know if you need anything else.